# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/extraction-queue/{id}/claim": {
      "post": {
        "operationId": "ExtractionQueueController_claim",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Atomically claim a candidate (queued → claimed)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/extraction-queue/{id}/accept": {
      "post": {
        "operationId": "ExtractionQueueController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Accept candidate and insert payer_rule",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/extraction-queue/{id}/reject": {
      "post": {
        "operationId": "ExtractionQueueController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reject candidate",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/extraction-queue/{id}/edit": {
      "post": {
        "operationId": "ExtractionQueueController_edit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Accept with edits and insert edited payer_rule",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/email-suppression": {
      "get": {
        "operationId": "SuppressionController_lookup",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Look up the suppression status of an address",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "SuppressionController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSuppressionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add an address to the global suppression list (admin override)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/email-suppression/{email}": {
      "delete": {
        "operationId": "SuppressionController_clear",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Clear suppression for an address (break-glass)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/cache/invalidate": {
      "post": {
        "operationId": "CacheInvalidateController_invalidate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvalidateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Bump the global synthesis-cache version (admin only)",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/rate-limit/overrides/{scope}": {
      "put": {
        "operationId": "RateLimitOverrideController_upsert",
        "parameters": [
          {
            "name": "scope",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertOverrideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Upsert an override for one scope",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "RateLimitOverrideController_remove",
        "parameters": [
          {
            "name": "scope",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove an override",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Billing Rules Platform API",
    "description": "Pre-flight a claim against payer × state × code rules.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id"
      }
    },
    "schemas": {
...
      "ClaimDto": {
        "type": "object",
        "properties": {}
      },
      "AcceptDto": {
        "type": "object",
        "properties": {}
      },
      "RejectDto": {
        "type": "object",
        "properties": {}
      },
      "EditDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSuppressionDto": {
        "type": "object",
        "properties": {}
      },
      "InvalidateDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpsertOverrideDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}